Add test modules for H2 and Testcontainers code generation
- Restructure project as Gradle composite build with root + included codegen module
- Add codegen-h2-test module for testing with in-memory H2 database
- Add codegen-tc-test module for testing with PostgreSQL via Testcontainers
- Update CI workflow to use composite build structure
- Migrate all modules to use Gradle version catalog
- Update documentation to reflect new multi-module architecture
- Remove obsolete template/entity.ftl file
