+++ .github/workflows/ci.yml
@@ -19,22 +19,11 @@ jobs:
           distribution: 'zulu'
           java-version: 21
 
-      - name: Grant execute permission for codegen/gradlew
-        working-directory: ./codegen
-        run: chmod +x gradlew
-
-      - name: Build
-        working-directory: ./codegen
-        run: ./gradlew build
-
       - name: Grant execute permission for ./gradlew
         run: chmod +x gradlew
 
-      - name: Test plugin (generate Java code)
-        run: ./gradlew clean domaCodeGenJavaAll build
-
-      - name: Test plugin (generate Kotlin code)
-        run: ./gradlew clean domaCodeGenKotlinAll build
+      - name: Build plugin
+        run: ./gradlew build
 
       - name: Set version
         id: set-version
@@ -45,5 +34,4 @@ jobs:
 
       - name: Publish plugin
         if: github.event_name == 'push' && endsWith(steps.set-version.outputs.version, 'SNAPSHOT') == false
-        working-directory: ./codegen
-        run: ./gradlew publishPlugins -Pgradle.publish.key=${{ secrets.GRADLE_PUBLISH_KEY }} -Pgradle.publish.secret=${{ secrets.GRADLE_PUBLISH_SECRET }}
+        run: ./gradlew :codegen:publishPlugins -Pgradle.publish.key=${{ secrets.GRADLE_PUBLISH_KEY }} -Pgradle.publish.secret=${{ secrets.GRADLE_PUBLISH_SECRET }}
