+++ codegen-h2-test/.gitignore
@@ -1,0 +1,19 @@
+/.classpath
+/.gradle/
+/.project
+/.settings/
+/bin/
+/build/
+/target/
+/.metadata
+/.idea/
+.factorypath
+/.claude/
+/data/
+/.apt_generated/
+
+/src/main/java/codegen
+/src/main/kotlin/codegen
+/src/main/resources/META-INF/codegen
+/src/test/java/codegen
+/src/test/kotlin/codegen
