+++ gradle.properties
@@ -1,4 +1,6 @@
 group=org.domaframework.doma
 version=3.2.1-SNAPSHOT
 release.useAutomaticVersion=true
-domaVersion=3.9.1
+org.gradle.parallel=true
+org.gradle.caching=true
+org.gradle.jvmargs=-Xmx2048M
