+++ gradle/libs.versions.toml
@@ -2,6 +2,7 @@
 # Core dependencies
 doma = "3.9.1"
 freemarker = "2.3.34"
+h2 = "2.3.232"
 postgresql = "42.7.7"
 
 # Testing
@@ -11,7 +12,6 @@ testcontainers = "1.21.2"
 # Plugins
 kotlin = "2.2.0"
 domaCompile = "4.0.0"
-runsql = "1.0.3"
 release = "3.1.0"
 spotless = "7.0.4"
 pluginPublish = "1.3.1"
@@ -28,6 +28,7 @@ doma-processor = { group = "org.seasar.doma", name = "doma-processor", version.r
 freemarker = { group = "org.freemarker", name = "freemarker", version.ref = "freemarker" }
 
 # Database drivers
+h2 = { group = "com.h2database", name = "h2", version.ref = "h2" }
 postgresql = { group = "org.postgresql", name = "postgresql", version.ref = "postgresql" }
 
 # Testing - BOMs
