+++ settings.gradle.kts
@@ -3,3 +3,6 @@ rootProject.name = "doma-codegen-plugin"
 pluginManagement {
     includeBuild("codegen")
 }
+
+include("codegen-h2-test")
+include("codegen-tc-test")
