--- template/entity.ftl
@@ -1,1 +1,0 @@
-hello
